BAT-43 - half-baked threads, needs fixing
